requirements.txt
openai==0.27.4
opencv_python==4.6.0.66
Pillow==10.0.1
PyPDF2==3.0.1
pytesseract==0.3.10
python-dotenv==1.0.0
requests==2.28.1
textract==1.6.5
wikipedia==1.4.0
